requirements.txt

joblib
fastapi
uvicorn
xgboost
pandas
numpy
plotly
streamlit
python-dotenv
redis
sendgrid
scikit-learn
matplotlib
seaborn
